add about swagger.yaml & chagne post cart response

+++ backend/swagger/model_store_info.go
@@ -9,7 +9,8 @@
 package swagger
 
 type StoreInfo struct {
-	Id int `json:"id"`
+
+	Id int64 `json:"id"`
 
 	Name string `json:"name"`
